+++ packages/terra-core-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Migrate docs from terra-search-field.
+
 ## 1.12.1 - (December 14, 2021)
 
 * Added
