+++ packages/terra-core-docs/src/terra-dev-site/doc/search-field/SearchField.1.doc.mdx
@@ -1,17 +1,17 @@
 import { Badge } from 'terra-search-field/package.json?dev-site-package';
 
-import SearchFieldBasic from '../example/SearchFieldBasic?dev-site-example';
-import SearchFieldPlaceholder from '../example/SearchFieldPlaceholder?dev-site-example';
-import SearchFieldMinimumLength from '../example/SearchFieldMinimumLength?dev-site-example';
-import SearchFieldDelayed from '../example/SearchFieldDelayed?dev-site-example';
-import SearchFieldDefaultValue from '../example/SearchFieldDefaultValue?dev-site-example';
-import SearchFieldDisabled from '../example/SearchFieldDisabled?dev-site-example';
-import SearchFieldDisabledAutoFormatting from '../example/SearchFieldDisableAutoFormatAttrs?dev-site-example';
-import SearchFieldBlock from '../example/SearchFieldBlock?dev-site-example';
-import SearchFieldDisableAutoSearch from '../example/SearchFieldDisableAutoSearch?dev-site-example';
-import SearchFieldFilterNumeric from '../example/SearchFieldFilterNumeric?dev-site-example';
-import SearchFieldFocus from '../example/SearchFieldFocus?dev-site-example';
-import SearchFieldProgrammaticSet from '../example/SearchFieldProgrammaticSet?dev-site-example';
+import SearchFieldBasic from './example/SearchFieldBasic?dev-site-example';
+import SearchFieldPlaceholder from './example/SearchFieldPlaceholder?dev-site-example';
+import SearchFieldMinimumLength from './example/SearchFieldMinimumLength?dev-site-example';
+import SearchFieldDelayed from './example/SearchFieldDelayed?dev-site-example';
+import SearchFieldDefaultValue from './example/SearchFieldDefaultValue?dev-site-example';
+import SearchFieldDisabled from './example/SearchFieldDisabled?dev-site-example';
+import SearchFieldDisabledAutoFormatting from './example/SearchFieldDisableAutoFormatAttrs?dev-site-example';
+import SearchFieldBlock from './example/SearchFieldBlock?dev-site-example';
+import SearchFieldDisableAutoSearch from './example/SearchFieldDisableAutoSearch?dev-site-example';
+import SearchFieldFilterNumeric from './example/SearchFieldFilterNumeric?dev-site-example';
+import SearchFieldFocus from './example/SearchFieldFocus?dev-site-example';
+import SearchFieldProgrammaticSet from './example/SearchFieldProgrammaticSet?dev-site-example';
 
 import SearchFieldPropsTable from 'terra-search-field/src/SearchField?dev-site-props-table';
