+++ packages/terra-core-docs/src/terra-dev-site/doc/search-field/example/SearchFieldPlaceholder.jsx
