+++ packages/terra-core-docs/src/terra-dev-site/doc/search-field/example/SearchFieldProgrammaticSet.jsx
@@ -1,8 +1,8 @@
 import React, { useState, useRef } from 'react';
 import Button from 'terra-button/lib/Button';
-import SearchField from '../../../SearchField';
+import SearchField from 'terra-search-field';
 
-const SettingValueSearchField = () => {
+const SearchFieldProgrammaticSet = () => {
   const [searchText, setSearchText] = useState('');
   const [message, setMessage] = useState('');
   const [text, setText] = useState('');
@@ -10,20 +10,18 @@ const SettingValueSearchField = () => {
 
   const setTextClick = () => {
     if (inputRef.current) {
-      setTimeout(() => {
-        inputRef.current.value = 'IpsumLorem';
-        let customEvent;
-        if (typeof (Event) === 'function') {
-          customEvent = new Event('input', {
-            bubbles: true,
-            cancelable: false,
-          });
-        } else {
-          customEvent = document.createEvent('Event');
-          customEvent.initEvent('input', true, false);
-        }
-        inputRef.current.dispatchEvent(customEvent);
-      }, 1000);
+      inputRef.current.value = 'IpsumLorem';
+      let customEvent;
+      if (typeof (Event) === 'function') {
+        customEvent = new Event('input', {
+          bubbles: true,
+          cancelable: false,
+        });
+      } else {
+        customEvent = document.createEvent('Event');
+        customEvent.initEvent('input', true, false);
+      }
+      inputRef.current.dispatchEvent(customEvent);
     }
   };
 
@@ -47,9 +45,9 @@ const SettingValueSearchField = () => {
 
   return (
     <div>
-      <div> Minimum Search Length is 3 </div>
+      <Button text="Set search field text" onClick={setTextClick} id="search-field-set-text-button" />
+      <p>Minimum Search Length is 3</p>
       <SearchField
-        inputRef
         id="searchfield"
         onSearch={handleSearch}
         onInvalidSearch={handleInvalidSearch}
@@ -59,7 +57,6 @@ const SettingValueSearchField = () => {
         value={text}
         inputRefCallback={(ref) => { inputRef.current = ref; }}
       />
-      <Button text="Set search field text" onClick={setTextClick} id="search-field-set-text-button" />
       <div id="search-callback-text">
         {message}
         {searchText}
@@ -68,4 +65,4 @@ const SettingValueSearchField = () => {
   );
 };
 
-export default SettingValueSearchField;
+export default SearchFieldProgrammaticSet;
