+++ packages/terra-core-docs/src/terra-dev-site/test/search-field/MinimumLengthSearchField.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SearchField from '../../../SearchField';
+import SearchField from 'terra-search-field';
 
 class MinimumLengthSearchField extends React.Component {
   constructor(props) {
