+++ packages/terra-core-docs/src/terra-dev-site/test/search-field/SearchFieldDisabled.test.jsx
@@ -1,4 +1,4 @@
 import React from 'react';
-import SearchField from '../../../SearchField';
+import SearchField from 'terra-search-field';
 
 export default () => <SearchField id="searchfield" isDisabled />;
