+++ packages/terra-core-docs/src/terra-dev-site/test/search-field/SearchFieldEnter.test.jsx
@@ -1,11 +1,12 @@
 import React from 'react';
-import SearchField from '../../../SearchField';
+import SearchField from 'terra-search-field';
 
 class AutoSearchDisabledSearchField extends React.Component {
   constructor(props) {
     super(props);
 
     this.state = {
+      searchCount: 0,
       searchText: '',
       message: '',
     };
@@ -15,26 +16,26 @@ class AutoSearchDisabledSearchField extends React.Component {
   }
 
   handleSearch(searchText) {
-    this.setState({ searchText, message: 'Search Text: ' });
+    this.setState(prevState => ({ searchCount: prevState.searchCount + 1, searchText, message: 'Search Text: ' }));
   }
 
   handleInvalidSearch(searchText) {
-    this.setState({ searchText, message: 'INVALID Search Text: ' });
+    this.setState(prevState => ({ searchCount: prevState.searchCount + 1, searchText, message: 'INVALID Search Text: ' }));
   }
 
   render() {
     return (
       <div>
-        <div> Auto Searching is Disabled </div>
-        <SearchField
-          id="searchfield"
-          onSearch={this.handleSearch}
-          onInvalidSearch={this.handleInvalidSearch}
-          disableAutoSearch
-        />
+        <h3> Auto Searching is Disabled </h3>
+        <SearchField id="searchfield" onSearch={this.handleSearch} onInvalidSearch={this.handleInvalidSearch} />
         <div id="search-callback-text">
           {this.state.message}
           {this.state.searchText}
+        </div>
+        <div id="search-count">
+          Search Count:
+          {' '}
+          {this.state.searchCount}
         </div>
       </div>
     );
