+++ packages/terra-core-docs/src/terra-dev-site/test/search-field/SearchFieldOnChange.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SearchField from '../../../SearchField';
+import SearchField from 'terra-search-field';
 
 class SearchFieldOnChange extends React.Component {
   constructor(props) {
